+++ GHEtool/VariableClasses/PipeData/Separatus.py
@@ -77,8 +77,7 @@ class Separatus(SingleUTube):
         -------
         Reynolds number : float
         """
-        u = flow_rate_data.mfr(fluid_data=fluid_data, **kwargs) / fluid_data.rho(**kwargs) / \
-            (pi * (0.02551 / 2) ** 2)
+        u = flow_rate_data.mfr(fluid_data=fluid_data, **kwargs) / fluid_data.rho(**kwargs) / (705.27 * 1e-6)
         return fluid_data.rho(**kwargs) * u * 0.02551 / fluid_data.mu(**kwargs)
 
     def pressure_drop(self, fluid_data: _FluidData, flow_rate_data: _FlowData, borehole_length: float,
@@ -109,7 +108,7 @@ class Separatus(SingleUTube):
             fluid_data.mu(**kwargs),
             fluid_data.rho(**kwargs),
             self.epsilon)
-        A = pi * (0.02551 / 2) ** 2
+        A = 705.27 * 1e-6  # cross-sectional area of the separatus
         V = (flow_rate_data.vfr(fluid_data=fluid_data, **kwargs) / 1000) / A
 
         # add 0.2 for the local losses