+++ CHANGELOG.md
@@ -7,14 +7,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [2.3.4] - Unpublished
 
-### Added
+###
 
+- Added Turbocollector from Muovitech.
 - Add flag in pressure drop calculation to in or exclude the pressure drop in the bend.
 - Added conical borehole heat exchangers (like the Vario probe from HakaGerodur).
 
 ### Fixed
 
 - Small bug in pressure drop lateral pipe.
+- Small bug in pressure drop calculation of separatus.
 
 ## [2.3.3] - 2025-05-28